22-10275 ConfirmationView components moved out of helpers file
