+++ src/applications/edu-benefits/10275/tests/containers/ConfirmationPage.unit.spec.jsx
@@ -6,7 +6,12 @@ import { render, cleanup } from '@testing-library/react';
 import { createInitialState } from '@department-of-veterans-affairs/platform-forms-system/state/helpers';
 
 import formConfig from '../../config/form';
-import ConfirmationPage from '../../containers/ConfirmationPage';
+import ConfirmationPage, {
+  ConfirmationGoBackLink,
+  ConfirmationPrintThisPage,
+  ConfirmationSubmissionAlert,
+  ConfirmationWhatsNextProcessList,
+} from '../../containers/ConfirmationPage';
 
 const mockStore = state => createStore(() => state);
 
@@ -31,6 +36,72 @@ const getPage = (data = {}, submission) =>
 
 describe('ConfirmationPage', () => {
   afterEach(cleanup);
+
+  describe('<ConfirmationSubmissionAlert />', () => {
+    it('shows process list section', () => {
+      const { getByText } = render(<ConfirmationSubmissionAlert />);
+
+      expect(getByText(/If we have any further questions/)).to.exist;
+    });
+  });
+
+  describe('<ConfirmationPrintThisPage />', () => {
+    it('should handle rendering summary box when no details are provided', () => {
+      const data = {
+        authorizedOfficial: {
+          fullName: {},
+        },
+      };
+      const submission = {};
+      const { getByTestId } = render(
+        <ConfirmationPrintThisPage data={data} submission={submission} />,
+      );
+
+      expect(getByTestId('full-name').innerHTML).to.contain('---');
+      expect(getByTestId('data-submitted').innerHTML).to.contain('---');
+    });
+
+    it('should render summary box with provided details', () => {
+      const data = {
+        authorizedOfficial: {
+          fullName: {
+            first: 'John',
+            middle: 'Test',
+            last: 'Doe',
+          },
+        },
+      };
+      const submitDate = new Date('09/18/2025');
+      const { getByTestId } = render(
+        <ConfirmationPrintThisPage data={data} submitDate={submitDate} />,
+      );
+
+      expect(getByTestId('full-name').innerHTML).to.contain('John Test Doe');
+      expect(getByTestId('data-submitted').innerHTML).to.contain(
+        'Sep 18, 2025',
+      );
+    });
+  });
+
+  describe('<ConfirmationWhatsNextProcessList />', () => {
+    it('shows process list section', () => {
+      const { getByText } = render(<ConfirmationWhatsNextProcessList />);
+
+      expect(getByText(/What to expect next/)).to.exist;
+      expect(getByText(/Your form will be evaluated/)).to.exist;
+    });
+  });
+
+  describe('<ConfirmationGoBackLink />', () => {
+    it('should render an action link to go back to the VA.gov homepage', () => {
+      const { container } = render(<ConfirmationGoBackLink />);
+
+      expect(container.querySelector('va-link-action')).to.have.attribute(
+        'text',
+        'Go back to VA.gov',
+      );
+    });
+  });
 
   it('shows success alert for "newCommitment" agreement type', () => {
     const { container, getByText } = getPage({
@@ -57,50 +128,23 @@ describe('ConfirmationPage', () => {
   });
 
   it('shows summary box with button to print page', () => {
-    const { container } = getPage({
-      response: {
-        attributes: {
-          confirmationNumber: '1234567890',
+    const { container } = getPage(
+      {},
+      {
+        response: {
+          attributes: {
+            confirmationNumber: '1234567890',
+          },
         },
+        timestamp: new Date().toISOString(),
       },
-      timestamp: new Date().toISOString(),
-    });
+    );
 
     expect(container.querySelector('va-summary-box')).to.exist;
     expect(container.querySelectorAll('va-summary-box h4').length).to.equal(3);
     expect(container.querySelector('va-button')).to.have.attribute(
       'text',
       'Print this page',
-    );
-  });
-
-  it('shows whats next section', () => {
-    const { getByText } = getPage({
-      response: {
-        attributes: {
-          confirmationNumber: '1234567890',
-        },
-      },
-      timestamp: new Date().toISOString(),
-    });
-
-    expect(getByText(/What to expect next/)).to.exist;
-    expect(getByText(/Your form will be evaluated/)).to.exist;
-  });
-
-  it('shows action link to return to VA.gov homepage', () => {
-    const { container } = getPage({
-      response: {
-        attributes: {
-          confirmationNumber: '1234567890',
-        },
-      },
-      timestamp: new Date().toISOString(),
-    });
-
-    expect(container.querySelector('va-link-action')).to.have.attribute(
-      'text',
-      'Go back to VA.gov',
     );
   });
 });
